util.py
import os
import sys
import yaml


def load_config():
    script_path = os.path.dirname(__file__)

    try:
        config_file = f'{script_path}/config.yml'

        with open(config_file, 'r') as stream:
            return yaml.safe_load(stream)
    except FileNotFoundError:
        print(f'ERROR: Config file {config_file} not found!')
        sys.exit()
