requirements.txt
requests==2.28.1
PyYAML==6.0
